+++ interfaces/Contracts/QueryInterface.php
@@ -17,20 +17,26 @@ use Countable;
 use Deprecated;
 use Iterator;
 use IteratorAggregate;
+use League\Uri\QueryBuildingMode;
 use Stringable;
 
 /**
  * @extends IteratorAggregate<array{0:string, 1:string|null}>
  *
+ * @method string|null toFormData() Returns the string representation using the application/www-form-urlencoded rules
+ * @method string|null toRFC3986() Returns the string representation using RFC3986 rules
+ * @method string|null first(string $key) returns the first value associated with the given name
+ * @method string|null last(string $key) returns the first value associated with the given name
+ * @method self normalize() returns the normalized string representation of the component
  * @method self withoutPairByKey(string ...$keys) Returns an instance without pairs with the specified keys.
  * @method self withoutPairByValue(Stringable|string|int|bool|null ...$values) Returns an instance without pairs with the specified values.
  * @method self withoutPairByKeyValue(string $key, Stringable|string|int|bool|null $value) Returns an instance without pairs with the specified key/value pair
  * @method bool hasPair(string $key, ?string $value) Tells whether the pair exists in the query.
- * @method string|null toFormData() Returns the string representation using the application/www-form-urlencoded rules
- * @method string|null toRFC3986() Returns the string representation using RFC3986 rules
- * @method self normalize() returns the normalized string representation of the component
- * @method string|null first(string $name) returns the first value associated with the given name
- * @method string|null last(string $name) returns the first value associated with the given name
+ * @method self appendParameterList(string $name, array $values, QueryBuildingMode $queryBuildingMode = QueryBuildingMode::Native) Appends a parameter to the query string
+ * @method self withParameterList(string $name, array $values, QueryBuildingMode $queryBuildingMode = QueryBuildingMode::Native) Adds a new parameter to the query string and remove any previously set values
+ * @method self withoutParameterList(string ...$names) Removes any given list associated with the given names
+ * @method array parameterList(string $name) Returns the list associated with the given name or an empty array if it does not exist.
+ * @method bool hasParameterList(string $name) Tells whether the parameter list exists in the query.
  */
 interface QueryInterface extends Countable, IteratorAggregate, UriComponentInterface
 {